Consolidate VertexAttribute code into the shader module.
