--- doc/modules/graphics/vertexattribute.rst
@@ -1,6 +1,0 @@
-pyglet.graphics.vertexattribute
-===============================
-
-.. automodule:: pyglet.graphics.vertexattribute
-  :members:
-  :undoc-members:
